requirements.txt
pyyaml
joblib
pandas
pyrender
roboticstoolbox-python
thop
torchnet
lxml
easydict
opencv-python==4.11.0.86
kornia
seaborn==0.11
numpy==1.24
matplotlib==3.7.3
pycollada
